Rename org.gradle.demos.* to org.gradle.xdcl.ecosystem.*
The built-in ecosystems outgrew their demo-era namespace. The mapping is
common -> org.gradle.xdcl.ecosystem.common, support ->
org.gradle.xdcl.ecosystem.support, java -> org.gradle.xdcl.ecosystem.java,
plugindev -> org.gradle.xdcl.ecosystem.plugindev; the schema packages move
with their libraries (.dsl suffix unchanged) and the carrier .xdcl files
now name the reactions by their new FQNs.

Giving xdcl-ecosystem-support its own .support package also un-splits
org.gradle.demos.common, which used to span the common-ecosystem and
ecosystem-support jars. org.gradle.xdcl.** sits outside PublicApi.includes
exactly as org.gradle.demos.** did, so the public-API standing of the
facades is unchanged. Remaining demo-era phrasing in docs goes with it.

The sibling checkout's integrations/gradle/doc/builtin-ecosystem-schemas.md
still cites org.gradle.demos and needs a follow-up there.
