+++ platforms/core-configuration/xdcl-common-ecosystem/src/main/xdcl/common.xdsl
@@ -1,8 +1,8 @@
-package org.gradle.demos.common.dsl
+package org.gradle.xdcl.ecosystem.common.dsl
 
 /// Library dependencies, mirroring the four JVM dependency scopes. Each is a list of coordinate
 /// notations (e.g. "group:name:version"). Shared so the wiring is declared once; every built-in
-/// ecosystem schema `import`s `org.gradle.demos.common.dsl`.
+/// ecosystem schema `import`s `org.gradle.xdcl.ecosystem.common.dsl`.
 record Dependencies {
   api: [String]?
   implementation: [String]?
@@ -11,7 +11,7 @@ record Dependencies {
 }
 
 /// A repository: a builtin marker symbol, or a String read as a maven repository URL. Declared under
-/// the demo's own names (not the settings bootstrap's) so the `:mavenCentral`/`:gradlePluginPortal`
+/// the ecosystem's own names (not the settings bootstrap's) so the `:mavenCentral`/`:gradlePluginPortal`
 /// symbols resolve to compilable facades in this package rather than colliding with the unshipped
 /// bootstrap types of the same simple name.
 type Repository = NamedRepository | String
@@ -20,7 +20,7 @@ type NamedRepository = Symbol where it in [:gradlePluginPortal, :mavenCentral]
 /// Capability trait for project types that configure project repositories. Composed by every JVM
 /// project-type template (`with … & HasRepositories`) so the repository surface is declared once and
 /// the wiring lives in common build logic that accepts this trait's facade (see
-/// `org.gradle.demos.common.Repositories`).
+/// `org.gradle.xdcl.ecosystem.support.Repositories`).
 trait HasRepositories {
   /// Repositories the reaction configures on the project so `dependencies` resolve.
   repositories: [Repository]?
@@ -29,7 +29,7 @@ trait HasRepositories {
 /// Capability trait for anything carrying library dependencies, by scope. Composed both at the
 /// project-type level (every JVM template is `with … & HasDependencies`) and per source set (each
 /// source record is `with HasDependencies`), so the shared scope/resolvable wiring in
-/// `org.gradle.demos.common.DependencyScopes` accepts either facade uniformly.
+/// `org.gradle.xdcl.ecosystem.support.DependencyScopes` accepts either facade uniformly.
 trait HasDependencies {
   /// Library dependencies, by scope.
   dependencies: Dependencies?
