+++ platforms/extensibility/xdcl-plugin-development-plugin/src/main/xdcl/plugin-development-ecosystem.xdcl
@@ -1,3 +1,3 @@
 plugin {
-  reactions [ :org.gradle.demos.plugindev.XdclGradlePluginReaction ]
+  reactions [ :org.gradle.xdcl.ecosystem.plugindev.XdclGradlePluginReaction ]
 }
