+++ platforms/jvm/xdcl-jvm-ecosystem-plugin/src/main/xdcl/java-ecosystem.xdcl
@@ -1,5 +1,5 @@
 plugin {
-  reactions [ :org.gradle.demos.java.JavaLibraryReaction ]
+  reactions [ :org.gradle.xdcl.ecosystem.java.JavaLibraryReaction ]
   defaults [
     JavaComponent {
       sources [ { name "main" }, { name "test" } ]
