--- platforms/jvm/xdcl-jvm-ecosystem/src/main/java/org/gradle/demos/java/JavaLibraryModel.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2026 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gradle.demos.java;
-
-import org.gradle.api.NamedDomainObjectContainer;
-import org.gradle.api.file.RegularFileProperty;
-import org.gradle.api.tasks.Nested;
-import org.gradle.demos.common.TestReports;
-
-/**
- * The imperative build-side state the {@link JavaLibraryReaction} produces and that other build logic
- * could consume.
- *
- * <p>Fully managed so {@code project.getExtensions().create("javaLibraryModel", JavaLibraryModel.class)}
- * instantiates it and auto-creates the {@link #getClasses()} container.
- */
-public interface JavaLibraryModel {
-
-    RegularFileProperty getJarFile();
-
-    NamedDomainObjectContainer<JavaClasses> getClasses();
-
-    @Nested
-    TestReports getTestReports();
-}
